Lib: proper size type in name_generator. Cleanup.

+++ lib/toolbox/name_generator.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -12,14 +13,14 @@ extern "C" {
  * @param      max_name_size  length of given buffer
  * @param[in]  prefix         The prefix of the name
  */
-void name_generator_make_auto(char* name, uint8_t max_name_size, const char* prefix);
+void name_generator_make_auto(char* name, size_t max_name_size, const char* prefix);
 
 /** Generates random name
  *
  * @param      name           buffer to write random name
  * @param      max_name_size  length of given buffer
  */
-void name_generator_make_random(char* name, uint8_t max_name_size);
+void name_generator_make_random(char* name, size_t max_name_size);
 
 /** Generates detailed name
  *
@@ -27,7 +28,7 @@ void name_generator_make_random(char* name, uint8_t max_name_size);
  * @param      max_name_size  length of given buffer
  * @param[in]  prefix         The prefix of the name
  */
-void name_generator_make_detailed(char* name, uint8_t max_name_size, const char* prefix);
+void name_generator_make_detailed(char* name, size_t max_name_size, const char* prefix);
 
 #ifdef __cplusplus
 }
